Set the linesize in VideoScaler to avoid padding
If linesize is unset, and the video size isn't aligned to a multiple of
16, then libav chooses an aligned number and leaves some bytes of
padding. This is a problem if you want to be able to copy the bytes
wholesale to e.g. the GPU, and it comes up for some video sizes like
2310x1080 (the 2310 becomes 2312 pixels across, where 2 are padding).

This change sets the linesize explictly during scaling (only if the
output format is RGBA) so that this padding is avoided.